Update libsynth.a to libsynth2.a
Overwrite the file in-place, since it's been
almost 20 years, and no one has reported any
issues with libsynth2 Vs. libsynth.

Retrieved from here:

http://www.dragonshadow.com/-/downloads/libsynth2.zip

On December 17th, 2023

Original notes describing the changes in this
version of libsynth:

Andreas kindly sent me an updated version of the
synth library. This fixes the incompatibility
between the Boot Loader and apps that use the
synth. Notably, Snake now works with the Boot
Loader when compiled with this updated library.

All new apps that use the synth should use this
library to avoid incompatibility with the Boot
Loader.