Fix certain single-precision floating point multiplies
Nuon doesn't include hardware floating point
instructions, and the soft-float implementation of
single-precision floating point multiplication in
libgcc.a was found to be buggy for certain values,
namely whenever rounding up the mantissa product
overflowed the mantissa, the overflow was ignored
and allowed to corrupt the exponent, causing
results to be either half or double what they were
supposed to be.

This change includes a fixed implementation of the
multiplication routine created by making minimal
modifications to the disassembled original
routine, as well as a script to build the new
version and patch it into libgcc.a, and a small
test program that exercises various problematic
and non-problematic cases and contains a C
transcription of the soft-float multiplication
algorithm used in the Nuon assembly for local
testing and algorithm experimentation.