Speed up the windows mgcc/mg++ wrappers
Keep it simple, stupid.

Perviously, these wrappers were invoking
Powershell to perform a search/replace operation
on the spec file similar to the sed operation on
the Linux wrappers that replaced %%VMLABSDIR%%
with the value in the VMLABS environment variable.
However, we don't need a fully general regex
search/replace operation. It's sufficient to just
insert the lines setting the vmlabsdir spec file
variable before concatenating on the rest of the
spec file template, which can be done trivially
with both standard Windows batch file commands and
the Unix echo+cat commands.

Starting a Powershell instance for each mgcc/mg++
invocation was taking forever, so this is a huge
speedup for Windows compilation.