Document ability to create NUON.CD and nuon.run files
Now that the signing tools are in the SDK, update
the README.md to note this, rather than telling
users it's coming in the future. Also better
document how to create NUON.CD files, and how to
burn NUON.CD and nuon.run files to discs that will
work across all Nuon systems.