Fix memory corruption in MGL BMP texture loading
There was a missing break in one case of a switch
statement in mglInitBMPTexture(). The result was
that when decoding BMP data into an e655-format
texture, the texture data was also decoded as a
eGRB655 texture in the memory just below the
actual texture, corrupting whatever was there.
Many times, that happened to be the prior texture
loaded, but for the first texture, it was who
knows what.